Added option to format visual selection
Works the same as plain :Neoformat when you are formatter a selection in a buffer

When you are formatting a subregion in a markdown file for example, then
you must use the following syntax.

:Neoformat! python

Note: the ! and filetype of the selection are required

Added tests for visual selection.
Using fstrings so travis config can be simplified.

https://github.com/sbdchd/neoformat/issues/30