Allowing passing of ranges directly to formatting tools
Neoformat supports formatting only part of a file, but it just passes
that chunk to formatting tools. A lot of tools need the context of the
surrounding lines, but support passing the requested range as an
argument.

This commit adds support for ranges by allowing certain variables to be
substituted in the arguments. In that case the 'range_mode' option should be
to indicate that the tool itself will handle the range.

Most tools use range_mode 1, which means that the tool takes the
entire buffer and a range as input, and outputs the entire buffer with
only that range formatted.

Some tools use range_mode 2, which means that the tool takes same input
as range_mode 1, but outputs only the range that was formatted.