Run formatters asynchronously in nvim
Unroll loop into recursion, and use jobstart if available to start a
job.

Those sleeps in tests are necessary, because file is not yet modified
after NeoFormat returns, sleep 1 seems to do the job on my machine, but
obviously might break a slower one.

I also added a sanity check to compare with the original snippet for
async job, it's skipped for synchronous